requirements.txt
click>=6.7,<=7.0
Jinja2>=2.9.6<=3.0.0
jsonschema>=2.6.0<=3.0.0
pytest>=3.1.3<=4.0.0
PyYAML>=3.12<=4.0
